+++ src/ralph/conf.py
@@ -211,6 +211,7 @@ class Settings(BaseSettings):
     LRS_RESTRICT_BY_AUTHORITY: bool = False
     LRS_RESTRICT_BY_SCOPES: bool = False
     LRS_PARTIAL_SUCCESS_DEFAULT: bool = False
+    LRS_ELASTICSEARCH_VALIDATE_KEYS: bool = True
     SENTRY_CLI_TRACES_SAMPLE_RATE: float = 1.0
     SENTRY_DSN: Optional[str] = None
     SENTRY_IGNORE_HEALTH_CHECKS: bool = False